Moved win32_compat.h includes one step down in the include depth.
Duplicated MAX_PATH into storage_constants.h, then replaced win32_compat.h by storage_constants.h where possible.
Introduced win32_compat.h at many of the lower include levels, mostly for DWORD/UCHAR/BYTE.

+++ include/OWORLD.h
@@ -64,7 +64,7 @@ public:
 	ZoomMatrix   *zoom_matrix;
 	Location     *loc_matrix;
 
-	DWORD	 		 next_scroll_time;		 // next scroll time
+	unsigned long	 		 next_scroll_time;		 // next scroll time
 
 	char			 scan_fire_x;				// cycle from 0 to SCAN_FIRE_DIST-1
 	char			 scan_fire_y;
